requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.4.0
numpy==2.1.1
opencv-python-headless==4.9.0.80
gunicorn==21.2.0
huggingface-hub==0.20.3
timm==0.9.12